Allowing for signup with h=:host parameter
