+++ js/app/models/UserCredentials.js
@@ -17,6 +17,7 @@
 define(function(require) {
   require('jquery.cookie');
   var api = require('util/api');
+  var url = require('util/url');
   var config = require('config');
   var Backbone = require('backbone');
   var ModelBase = require('models/ModelBase');
@@ -70,22 +71,6 @@ define(function(require) {
       }
     },
     
-    _getQueryVariable: function(variable) {
-      var search = Backbone.history.location.search;
-      if (!search || search == '') {
-        return undefined;
-      }
-      var query = search.substring(1);
-      var vars = query.split("&");
-      for (var i = 0; i < vars.length; i++) {
-        var pair = vars[i].split("=");
-        if (pair[0] == variable) {
-          return unescape(pair[1]);
-        }
-      }
-      return undefined;
-    },
-
     isPermanent: function() {
       return Boolean(localStorage.loginPermanent);
     },
@@ -106,14 +91,7 @@ define(function(require) {
       var password = attributes.password;
 
       if (username && password) {
-        if (username.indexOf('@') == -1) {
-          if (config.useURLHostAsDomain) {
-            username += '@' + (this._getQueryVariable('h') || 
-                Backbone.history.location.hostname);
-          } else {
-            username += '@' + config.homeDomain;
-          }
-        }
+        username = this.normalizeJid(username);
 
         var credentials = btoa(username + ':' + password);
         this.set({'username': username, 'credentials': credentials});
@@ -122,7 +100,19 @@ define(function(require) {
         this._persist(options.permanent);
       }
     },
-
+    
+    normalizeJid: function(username) {
+      if (username.indexOf('@') == -1) {
+        if (config.useURLHostAsDomain) {
+          username += '@' + (url.getQueryVariable('h') || 
+              Backbone.history.location.hostname);
+        } else {
+          username += '@' + config.homeDomain;
+        }
+      }
+      return username;
+    },
+    
     addAuthorizationToAjaxOptions: function(options) {
       if (options) {
         options.headers = options.headers || {};
