+++ js/app/util/url.js
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2012 buddycloud
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+
+define(function(require) {
+  var Backbone = require('backbone');
+
+  function getQueryVariable(variable) {
+    var search = Backbone.history.location.search;
+    if (!search || search == '') {
+      return undefined;
+    }
+    var query = search.substring(1);
+    var vars = query.split("&");
+    for (var i = 0; i < vars.length; i++) {
+      var pair = vars[i].split("=");
+      if (pair[0] == variable) {
+        return unescape(pair[1]);
+      }
+    }
+    return undefined;
+  }
+
+  return {
+    getQueryVariable: getQueryVariable
+  };
+});
